% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{attitude_gap_analysis}
\alias{attitude_gap_analysis}
\title{Analyze Significance of Attitude Gaps}
\usage{
attitude_gap_analysis(
  data,
  issue,
  years = 1945:2022,
  grouping_var = c("gender", "degree", "union_household", "community_size"),
  ...
)
}
\arguments{
\item{data}{Data from inst/extdata/issue_data_with_weights.RDS.}

\item{issue}{Character string giving the issue to be used.}

\item{years}{Select desired years}

\item{grouping_var}{For which group do you want separate estimates (gender, degree, union_household, community_size)}

\item{...}{Other arguments to be passed down, not implemented}
}
\description{
Analyze Significance of Attitude Gaps
}
\examples{
iss <- readRDS(system.file("extdata/issue_data_with_weight.RDS", package="cvpa"))
attitude_gap_analysis(iss, "wageprice_1", grouping_var="gender")

}
