% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_flag}
\alias{make_flag}
\title{Make Reliability Flag for Weighted Proportions Test}
\usage{
make_flag(x, n)
}
\arguments{
\item{x}{A vector of two values of category counts.}

\item{n}{A vector of two values of sample sizes}
}
\description{
Makes a flag to identify if the p-value for the difference
of proportions test is reliable (all expected counts greater than 5).
}
