% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_aov}
\alias{sum_aov}
\title{Summarise aov object}
\usage{
sum_aov(x, ...)
}
\arguments{
\item{x}{An object of class `aov`.}

\item{...}{Other arguments to be passed down, currently not implemented.}
}
\description{
Summarise an anova object as a weighted difference
of means test
}
\examples{
data(issue_data)
a <- aov(val ~ union_household, 
         data=subset(issue_data, issue=="wageprice_1"))
sum_aov(a)

}
