% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vote_data}
\alias{vote_data}
\title{Vote Choice/Intention Data}
\format{
A data frame with 78,614 rows and 11 variables
\describe{
  \item{vote}{Factor indicating vote choice or intention depending on \code{type}}
  \item{type}{String indicating vote whever \code{vote} is intention or actual choice}
  \item{year}{Year of survey}
  \item{region}{Region of residence: 1 = Atlantic, 2 = Quebec, 3 = Ontario, 4 = Prairies, 5 = BC, 6 = The North}
  \item{province}{Province of residence: 10 = Newfoundland and Labrador, 11 = PEI, 12 = Nova Scotia, 13 = New Brunswick, 
  24 = Quebec, 35 = Ontario, 46 = Manitoba, 47 = Saskatchewan, 48 = Alberta, 59 = BC, 
  60 = Yukon, 61 = Northwest Territories, 62 = Nunavut}
  \item{community_size}{Community size >= 100,000 (0/1)}
  \item{gender}{Respondent gender 0 = male, 1 = female.  Non-binary category begins in 2019 and is available in the full data on the \href{https://www.queensu.ca/cora/}{Canadian Opinion Research Archive}.}
  \item{age_cats}{Age categories: 1 = 18-29, 2 = 30-49, 3 = 50+}
  \item{religion}{Add religion description}
  \item{degree}{Respondent has a university degree (0/1)}
  \item{union_household}{Whether respondent is in a household with a union member (0/1)}
  \item{language}{Add language description}
  \item{occupation}{Occupation: 1=Managers, Executives, Business Owners, 2=Professional, 3=Routine Non-Manual, 4=Working Class, 5=Armed Forces, 6=Farmer/Fisherman, 7=Retired, 8=Unemployed, 9=Student, 99=Other (incl. housewife, homemaker)}
  \item{weight}{Sum of weights for \code{region}-\code{langauge} profile}
}
}
\description{
Data on vote choice and intention from several surveys. The data are in long form (in intentnion/choice) and aggregate to profile across the variables `region`-`language` described below.
}
\details{
Fill in some details
}
\references{
Lucas, Jack, Tyler Romualdi, David A. Armstrong II and Elizabeth Goodyear-Grant. "Demographic Divides in Canadian Voting, 1945-2023: The Canadian Vote Intention Dataset", Canadian Journal of Political Science, Forthcoming.
}
\keyword{datasets}
