% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wtd_issue}
\alias{wtd_issue}
\title{Make weighted mean of vote intention/choice}
\usage{
wtd_issue(data, issue, years = 1945:2022, grouping_vars = NULL, ...)
}
\arguments{
\item{data}{Data from inst/extdata/issue_data_with_weights.RDS.}

\item{issue}{Character string giving name of issue variable}

\item{years}{Select desired years}

\item{grouping_vars}{For which groups do you want separate estimates (region or province, gender, age_cats, degree, religion, language, union_household, community_size)}

\item{...}{Other arguments to be passed down, not implemented}
}
\description{
Make weighted mean of vote intention/choice
}
\examples{
data(issue_data)
wtd_issue(issue_data, "wageprice_1", grouping_vars="degree")

}
