% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gap_analysis}
\alias{gap_analysis}
\title{Analyze Significance of Voting Gaps}
\usage{
gap_analysis(
  data,
  vote_type = c("intention", "choice"),
  incl_undecided = FALSE,
  years = 1945:2022,
  grouping_var = c("age_cats", "religion", "degree", "gender", "province", "region",
    "language", "union_household", "community_size", "occupation"),
  levels = NULL,
  ...
)
}
\arguments{
\item{data}{Data from inst/extdata/integrated_with_weights.RDS.}

\item{vote_type}{Use vote intention or choice}

\item{incl_undecided}{If using vote intention, should undecided voters be included?}

\item{years}{Select desired years}

\item{grouping_var}{For which group do you want separate estimates (gender, degree, union_household, community_size)}

\item{levels}{Character vector of length 2 giving values of `grouping_var` to use for comparison.  `NULL` is valid if `grouping_var` only has two valid values.  Otherwise, values must be specified.}

\item{...}{Other arguments to be passed down, not implemented}
}
\description{
The function does a difference of proportions test for the
two identified groups for each of the parties.  Note, for gender,
there are not many surveys with non-binary/other response options
so those are removed from consideration here.
}
\details{
See [response_gap_analysis()] for the same test generalized to
  any policy question, not just vote choice/intention.
}
\examples{
data(vote_data)
gap_analysis(vote_data, "intention", years=2000:2022, grouping_var="gender")

}
