% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gap_analysis}
\alias{gap_analysis}
\title{Analyze Significance of Voting Gaps}
\usage{
gap_analysis(
  data,
  vote_type = c("intention", "choice"),
  incl_undecided = FALSE,
  years = 1945:2022,
  grouping_var = c("gender", "degree", "union_household", "community_size"),
  ...
)
}
\arguments{
\item{data}{Data from inst/extdata/integrated_with_weights.RDS.}

\item{vote_type}{Use vote intention or choice}

\item{incl_undecided}{If using vote intention, should undecided voters be included?}

\item{years}{Select desired years}

\item{grouping_var}{For which group do you want separate estimates (gender, degree, union_household, community_size)}

\item{...}{Other arguments to be passed down, not implemented}
}
\description{
Analyze Significance of Voting Gaps
}
\examples{
data(vote_data)
gap_analysis(vote_data, "intention", years=2000:2022, grouping_var="gender")

}
