% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wtd_vote}
\alias{wtd_vote}
\title{Make weighted mean of vote intention/choice}
\usage{
wtd_vote(
  data,
  vote_type = c("intention", "choice"),
  incl_undecided = FALSE,
  years = 1945:2022,
  grouping_vars = NULL,
  ...
)
}
\arguments{
\item{data}{Data from inst/extdata/integrated_with_weights.RDS.}

\item{vote_type}{Use vote intention or choice}

\item{incl_undecided}{If using vote intention, should undecided voters be included?}

\item{years}{Select desired years}

\item{grouping_vars}{For which groups do you want separate estimates (region or province, gender, age_cats, degree, religion, language, union_household, community_size)}

\item{...}{Other arguments to be passed down, not implemented}
}
\description{
Make weighted mean of vote intention/choice
}
\examples{
data(vote_data)
wtd_vote(vote_data, "intention", years=2000:2022)

}
